+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/RuntimeHelpers.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 
-namespace System.Runtime.CompilerServicesOverride
+namespace Microsoft.Data.SqlClient
 {
     internal struct RuntimeHelpers
     {
