+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SqlDataReader.cs
@@ -14,7 +14,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.IO;
 using System.Reflection;
-using RuntimeHelpers = System.Runtime.CompilerServicesOverride.RuntimeHelpers;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.Json;
