+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParserStateObject.netcore.cs
@@ -5,7 +5,6 @@
 using System;
 using System.Diagnostics;
 using System.Buffers.Binary;
-using System.Runtime.CompilerServicesOverride;
 using System.Runtime.InteropServices;
 using System.Security;
 using System.Threading;
