+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/TdsParserStateObject.cs
@@ -17,8 +17,6 @@ namespace Microsoft.Data.SqlClient
 #if NETFRAMEWORK
     using PacketHandle = IntPtr;
     using RuntimeHelpers = System.Runtime.CompilerServices.RuntimeHelpers;
-#else
-    using System.Runtime.CompilerServicesOverride;
 #endif
     
     sealed internal class LastIOTimer
